feat: udistrital/sga_documentacion#180 Refactorización del endpoint PostBaseLegalizacionMatricula
